Fix pvs issue on maps with far too many leafs (was still using 16bit leaf indexes, whoops).
git-svn-id: https://svn.code.sf.net/p/fteqw/code/trunk@5767 fc73d0e0-1445-4013-8a0c-d673dee63da5

+++ engine/common/bspfile.h
@@ -917,7 +917,7 @@ typedef struct
 typedef struct pvscache_s
 {
 	int				num_leafs;
-	unsigned short	leafnums[MAX_ENT_LEAFS];
+	unsigned int	leafnums[MAX_ENT_LEAFS];
 #if defined(Q2BSPS) || defined(Q3BSPS) || defined(TERRAIN)
 	int				areanum;	//q2bsp
 	int				areanum2;	//q2bsp
